requirements.txt
# Core dependencies for data collection
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0

# ESIOS API client
# Note: We'll implement our own client for better control

# Data storage
sqlalchemy>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
